requirements.txt
requests
pandas
numpy
gspread
gspread-dataframe
google-auth